tenant id added for eval models

+++ paig-server/backend/paig/api/evaluation/services/eval_target_service.py
@@ -1,12 +1,12 @@
+import copy
 import json
 import traceback
 from core.utils import SingletonDepends, is_valid_url
 from api.evaluation.database.db_operations.eval_target_repository import EvaluationTargetRepository
 import logging
-from api.governance.database.db_operations.ai_app_repository import AIAppRepository
 from core.exceptions import NotFoundException, InternalServerError, BadRequestException
 from core.controllers.paginated_response import create_pageable_response
-
+from core.db_session.transactional import Transactional, Propagation
 
 logger = logging.getLogger(__name__)
 
@@ -40,23 +40,21 @@ def transform_eval_target(eval_target):
 
 def transform_eval_target_to_dict(config):
     target = json.loads(config)
-    config = target['config']
+    config = target.get('config', {})
     eval_taget_config = dict()
-    eval_taget_config['method'] = config['method']
-    eval_taget_config['headers'] = config['headers']
-    eval_taget_config['body'] = config['body']
-    eval_taget_config['transformResponse'] = config['transformResponse']
-    eval_taget_config['url'] = target['id']
+    eval_taget_config['method'] = config.get('method', '')
+    eval_taget_config['headers'] = config.get('headers', {})
+    eval_taget_config['body'] = config.get('body', {})
+    eval_taget_config['transformResponse'] = config.get('transformResponse', '')
+    eval_taget_config['url'] = target.get('id', '')
     return eval_taget_config
 
 class EvaluationTargetService:
 
     def __init__(self,
-        eval_target_repository: EvaluationTargetRepository = SingletonDepends(EvaluationTargetRepository),
-        ai_app_repository: AIAppRepository = SingletonDepends(AIAppRepository)
+        eval_target_repository: EvaluationTargetRepository = SingletonDepends(EvaluationTargetRepository)
     ):
         self.eval_target_repository = eval_target_repository
-        self.ai_app_repository = ai_app_repository
 
 
     async def get_all_ai_app_with_host(self, include_filters, exclude_filters, page_number, size, sort):
@@ -72,20 +70,15 @@ class EvaluationTargetService:
         if ai_apps is None:
             raise NotFoundException("No applications found")
         try:
-            index = 1
             final_apps = list()
             for ai_app in ai_apps:
                 app = dict()
-                app['id'] = index
-                index += 1
-                app['ai_application_id'] = ai_app[0]
-                app['target_id'] = ai_app[1]
-                app['desc'] = ai_app[3]
-                app['url'] = ai_app[4]
-                if app['ai_application_id'] is not None:
-                    app['name'] = ai_app[2]
-                else:
-                    app['name'] = ai_app[5]
+                app['id'] = ai_app.id
+                app['ai_application_id'] = ai_app.application_id
+                app['target_id'] = ai_app.id
+                app['url'] = ai_app.url
+                app['name'] = ai_app.name
+                app['status'] = ai_app.status
                 final_apps.append(app)
             return create_pageable_response(final_apps, total_count, page_number, size, sort)
         except Exception as e:
@@ -93,46 +86,41 @@ class EvaluationTargetService:
             logger.error(traceback.format_exc())
             raise InternalServerError("Internal server error")
 
+    @Transactional(propagation=Propagation.REQUIRED)
     async def create_app_target(self, body_params):
         new_params = dict()
-        if 'ai_application_id' in body_params and body_params['ai_application_id']:
-            app_id = body_params['ai_application_id']
-            ai_app = await self.ai_app_repository.get_record_by_id(app_id)
-            if ai_app is None:
-                raise NotFoundException(f"No AI application found with id {app_id}")
-            # check for existing target
-            target_model = await self.eval_target_repository.get_target_by_app_id(app_id)
-            if target_model is not None:
-                raise BadRequestException(f"Target already exists for AI application {ai_app.name}")
-            body_params['name'] = ai_app.name
-            new_params['application_id'] = app_id
-        else:
-            name_exists = await self.eval_target_repository.application_name_exists(body_params['name'])
-            if name_exists:
-                raise BadRequestException(f"Application with name {body_params['name']} already exists")
+
         transformed_eval = transform_eval_target(body_params)
         new_params['config'] = transformed_eval
         new_params['name'] = body_params['name']
         new_params['url'] = body_params['url']
         new_params['target_user'] = body_params['username']
-        try:
+        new_params['status'] = 'ACTIVE'
+        if 'ai_application_id' in body_params and body_params['ai_application_id']:
+            app_id = body_params['ai_application_id']
+            # check for existing target
+            target_model = await self.eval_target_repository.get_target_by_app_id(app_id)
+            if target_model is None:
+                raise BadRequestException(f"AI application does not exists for give AI application id {app_id}")
+            body_params['name'] = target_model.name
+            new_params['application_id'] = app_id
+            eval_target = await self.eval_target_repository.update_app_target(new_params, target_model)
+            return eval_target
+        else:
+            name_exists = await self.eval_target_repository.application_name_exists(body_params['name'])
+            if name_exists:
+                raise BadRequestException(f"Application with name {body_params['name']} already exists")
             eval_target = await self.eval_target_repository.create_app_target(new_params)
             return eval_target
-        except Exception as e:
-            logger.error(f"Error in create_app_target: {e}")
-            logger.error(traceback.format_exc())
-            raise InternalServerError("Internal server error")
 
+    @Transactional(propagation=Propagation.REQUIRED)
     async def update_app_target(self, target_id, body_params):
         new_params = dict()
         target_model = await self.eval_target_repository.get_target_by_id(target_id)
         if target_model is None:
             raise NotFoundException(f"No  application found with id {target_id}")
         if target_model.application_id:
-            ai_app = await self.ai_app_repository.get_record_by_id(target_model.application_id)
-            if ai_app is None:
-                raise NotFoundException(f"No AI application found with id {target_model.application_id}")
-            body_params['name'] = ai_app.name
+            body_params['name'] = target_model.name
         else:
             if 'name' in body_params and body_params['name'] != target_model.name:
                 name_exists = await self.eval_target_repository.application_name_exists(body_params['name'])
@@ -141,26 +129,36 @@ class EvaluationTargetService:
         transformed_eval = transform_eval_target(body_params)
         new_params['config'] = transformed_eval
         new_params['url'] = body_params['url']
+        new_params['status'] = 'ACTIVE'
         if 'name' in body_params:
             new_params['name'] = body_params['name']
         if 'username' in body_params:
             new_params['target_user'] = body_params['username']
         try:
             eval_target = await self.eval_target_repository.update_app_target(new_params, target_model)
-            eval_target = eval_target.__dict__
-            eval_target['username'] = eval_target.pop('target_user')
-            return eval_target
+            eval_target_resp = copy.deepcopy(eval_target.__dict__)
+            eval_target_resp['username'] = eval_target_resp.pop('target_user')
+            return eval_target_resp
         except Exception as e:
             logger.error(f"Error in update_app_target: {e}")
             logger.error(traceback.format_exc())
             raise InternalServerError("Internal server error")
 
+    @Transactional(propagation=Propagation.REQUIRED)
     async def delete_target(self, app_id):
         target_model = await self.eval_target_repository.get_target_by_id(app_id)
         if target_model is None:
             raise NotFoundException(f"No application found with id {app_id}")
         try:
-            eval_target = await self.eval_target_repository.delete_target(target_model)
+            if target_model.application_id:
+                new_params = dict()
+                new_params['url'] = None
+                new_params['config'] = "{}"
+                new_params['status'] = 'INACTIVE'
+                new_params['target_user'] = "NA"
+                eval_target = await self.eval_target_repository.update_app_target(new_params, target_model)
+            else:
+                eval_target = await self.eval_target_repository.delete_target(target_model)
             return eval_target
         except Exception as e:
             logger.error(f"Error in delete_target: {e}")
@@ -180,8 +178,43 @@ class EvaluationTargetService:
             resp['target_id'] = target_model.id
             resp['ai_application_id'] = target_model.application_id
             resp['username'] = target_model.target_user
+            resp['status'] = target_model.status
             return resp
         except Exception as e:
             logger.error(f"Error in get_app_target_by_id: {e}")
+            logger.error(traceback.format_exc())
+            raise InternalServerError("Internal server error")
+
+    @Transactional(propagation=Propagation.REQUIRED)
+    async def update_app_target_on_ai_app(self, ai_app_id, ai_app_name):
+        # check for existing target
+        target_model = await self.eval_target_repository.get_target_by_app_id(ai_app_id)
+        if target_model is not None:
+            target_model.name = ai_app_name
+            return await self.eval_target_repository.update_app_target({'name': ai_app_name}, target_model)
+        new_params = dict()
+        new_params['application_id'] = ai_app_id
+        new_params['name'] = ai_app_name
+        new_params['url'] = None
+        new_params['config'] = "{}"
+        try:
+            eval_target = await self.eval_target_repository.create_app_target(new_params)
+            return eval_target
+        except Exception as e:
+            logger.error(f"Error in insert_app_target_on_ai_app: {e}")
+            logger.error(traceback.format_exc())
+            raise InternalServerError("Internal server error")
+
+    @Transactional(propagation=Propagation.REQUIRED)
+    async def delete_app_target_on_ai_app(self, ai_app_id):
+        # check for existing target
+        target_model = await self.eval_target_repository.get_target_by_app_id(ai_app_id)
+        if target_model is None:
+            return
+        try:
+            eval_target = await self.eval_target_repository.delete_target(target_model)
+            return eval_target
+        except Exception as e:
+            logger.error(f"Error in delete_app_target_on_ai_app: {e}")
             logger.error(traceback.format_exc())
             raise InternalServerError("Internal server error")
